Add more ways to disable telemetry
