+++ .golangci-new.yml
@@ -10,7 +10,15 @@ linters-settings:
     errorf: false
     asserts: true
     comparison: true
-
+  revive:
+    ignore-generated-header: true
+    severity: warning
+    # TODO enable-all-rules: true
+    rules:
+      # https://github.com/mgechev/revive/blob/master/RULES_DESCRIPTIONS.md
+      - name: exported
+        arguments: [checkPrivateReceivers]
+      - name: package-comments
   staticcheck:
     checks: ["all"]
   wsl:
@@ -48,7 +56,6 @@ linters:
     - lll
     - misspell
     - nolintlint
-    - revive
     - unused
     - whitespace
