+++ internal/util/debug/debug.go
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package debug provides debug facilities.
 package debug
 
 import (
