+++ internal/util/state/state.go
@@ -30,8 +30,9 @@ type State struct {
 	Telemetry *bool  `json:"telemetry,omitempty"` // nil for undecided
 
 	// never persisted
-	Start         time.Time `json:"-"`
-	LatestVersion string    `json:"-"`
+	Start           time.Time `json:"-"`
+	LatestVersion   string    `json:"-"`
+	TelemetryLocked bool      `json:"-"`
 }
 
 // fill replaces all unset or invalid values with default.
@@ -53,9 +54,10 @@ func (s *State) deepCopy() *State {
 	}
 
 	return &State{
-		UUID:          s.UUID,
-		Telemetry:     telemetry,
-		Start:         s.Start,
-		LatestVersion: s.LatestVersion,
+		UUID:            s.UUID,
+		Telemetry:       telemetry,
+		Start:           s.Start,
+		LatestVersion:   s.LatestVersion,
+		TelemetryLocked: s.TelemetryLocked,
 	}
 }
